+++ lib/lite/uxid/nanoid.rb
@@ -5,9 +5,11 @@ module Lite
     class Nanoid < Base::Irreversible
 
       def encode
-        (0...coder_size).each_with_object(+"") do |i, str|
+        uxid = (0...coder_size).each_with_object(+"") do |i, str|
           str << coder_charset[coder_bytes[i] & 63]
         end
+
+        "#{coder_prefix}#{uxid}"
       end
 
     end