+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.5.0] - 2024-09-20
 ### Added
 - Added uuid version option
+- Added prefix option to hashid and nanoid
 ### Changed
+- Move reversible and irreversible files to base namespace
 - Ruby send perf improvements
 
 ## [1.4.0] - 2022-11-20